Adding a validation so a user can only delete their own activities

+++ src/main/java/com/example/todo/service/ActivityService.java
@@ -3,12 +3,14 @@ package com.example.todo.service;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import javax.validation.constraints.NotNull;
 
 import org.springframework.stereotype.Service;
 
+import com.example.todo.exception.IllegalActivityDeletionAttempt;
 import com.example.todo.model.Activity;
 import com.example.todo.repository.ActivityRepository;
 
@@ -31,12 +33,24 @@ public class ActivityService {
         return activityRepository.save(activity);
     }
 
-    public void deleteById(UUID id) {
-        activityRepository.deleteById(id);
+    public void deleteById(UUID id, UUID user_id) throws IllegalActivityDeletionAttempt {
+    	
+    	Activity activity = findByIdAndFetchUser(id).get();
+ 
+    	// An activity can be deleted only by its owner
+    	if (!activity.getUser().getId().equals(user_id)) {
+    		throw new IllegalActivityDeletionAttempt("Trying to delete an activity from another user");
+    	}
+ 
+    	activityRepository.deleteById(id);
     }
 
 	public List<Activity> findByUserIdOrderByCreatedAtAsc(UUID userId) {
 		return activityRepository.findByUserIdOrderByCreatedAtAsc(userId);
+	}
+	
+	public Optional<Activity> findByIdAndFetchUser(UUID id) {
+		return activityRepository.findByIdAndFetchUser(id);
 	}
 
 }
